backend/requirements_full.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
sqlalchemy==2.0.27
asyncpg==0.29.0
pgvector==0.2.5
sentence-transformers>=2.5.1
torch>=2.0.0
transformers>=4.38.0
numpy>=1.26.4
openai==1.12.0
httpx==0.27.0
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langchain-openai==0.0.7
chromadb==0.4.22
apscheduler==3.10.4
pyyaml==6.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
# TTS (Coqui XTTS) removed - requires Python <3.12. Uses OpenAI TTS fallback instead.
openai-whisper>=20231117
pytesseract>=0.3.10
pymupdf>=1.23.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
soundfile==0.12.1
google-api-python-client==2.120.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
aiofiles==23.2.1
tenacity==8.2.3
minio==7.2.4
boto3==1.34.44
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
beautifulsoup4==4.12.3
requests==2.31.0
yt-dlp>=2024.3.10
docx2txt==0.8
pypdf==4.0.2
alembic==1.13.1
